+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Lastly, a new [release](https://github.com/MAKENTNU/web/releases) must be create
 ### Other changes
 
 - Added custom front page logo on [makentnu.dev](https://makentnu.dev/) (MAKENTNU/web#781)
+- Stopped generating `#:` comments in `.po` files (MAKENTNU/web#806)
 
 
 ## 2025-10-26 (MAKENTNU/web#779)